+++ src/common/styleUtils/keyframes.ts
@@ -8,3 +8,14 @@ export const spin = keyframes`
     transform: rotate(360deg);
     }
 `;
+
+export const slideDown = keyframes`
+  0% {
+    opacity: 0;
+    transform: translateY(-1.6rem);
+  }
+  100% {
+    opacity: 1;
+    transform: translateY(0);
+  }
+`;