+++ src/components/components/Modal/Styled.ts
@@ -1,11 +1,12 @@
 import styled from "styled-components";
 import { ModalContainerStyleProps } from "./types";
+import { slideDown } from "../../../common/styleUtils/keyframes";
 
 export const WModal = styled.div`
   position: fixed;
   display: flex;
-  padding: ${props => props.theme.space[0]};
-  overflow-y: scroll;
+  padding: ${props => `${props.theme.space[0]}px`};
+  overflow-y: auto;
   align-items: center;
   top: 0;
   bottom: 0;
@@ -26,7 +27,7 @@ export const ModalContainer = styled.div<ModalContainerStyleProps>`
   box-shadow: ${props => props.theme.shadows[1]};
   display: flex;
   flex-direction: column;
-  padding: 0 ${props => props.theme.space[2]};
+  padding: 0 ${props => `${props.theme.space[2]}px`};
   position: absolute;
   top: 0;
   margin: 1rem auto;
@@ -42,14 +43,14 @@ export const ModalContainer = styled.div<ModalContainerStyleProps>`
 
   width: 100%;
   z-index: 1;
-  animation: slide-down 0.2s ease 1;
+  animation: ${slideDown} 0.15s ease 1;
   will-change: transform;
 `;
 ModalContainer.displayName = "ModalContainer";
 
 export const ModalBody = styled.div`
   overflow-y: auto;
-  padding: ${props => props.theme.space[3]};
+  padding: ${props => `${props.theme.space[3]}px`};
   position: relative;
   margin-bottom: 1rem;
   max-width: 960px;
@@ -67,13 +68,13 @@ export const ModalOverlay = styled.div`
 ModalOverlay.displayName = "ModalOverlay";
 
 export const ModalHeader = styled.div`
-  padding: ${props => props.theme.space[3]};
-  margin-top: ${props => props.theme.space[2]};
+  padding: ${props => `${props.theme.space[3]}px`};
+  margin-top: ${props => `${props.theme.space[2]}px`};
 `;
 ModalHeader.displayName = "ModalHeader";
 
 export const ModalFooter = styled.div`
-  padding: ${props => props.theme.space[3]};
+  padding: ${props => `${props.theme.space[3]}px`};
   text-align: right;
 `;
 ModalFooter.displayName = "ModalFooter";