+++ CHANGELOG.md
@@ -15,6 +15,7 @@ All notable changes to this project will be documented in this file.
 - **InputField**
 - **InputForm**
 - **FormGroup**
+- Update **Button** auto focus style
 
 ### Breaking Change